+++ JLMoviesWizeL/.idea/gradle.xml
@@ -11,6 +11,7 @@
           <set>
             <option value="$PROJECT_DIR$" />
             <option value="$PROJECT_DIR$/app" />
+            <option value="$PROJECT_DIR$/app/databasemovies" />
           </set>
         </option>
       </GradleProjectSettings>
