+++ JLMoviesWizeL/settings.gradle
@@ -14,3 +14,4 @@ dependencyResolutionManagement {
 }
 rootProject.name = "JLMovies WizeL"
 include ':app'
+include ':app:databasemovies'
